web routes with mw

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\ListingController;
+use App\Http\Controllers\UserController;
 use App\Models\Listing;
 use Illuminate\Support\Facades\Route;
 
@@ -16,9 +17,14 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::get('/', [ListingController::class, 'index']);
-Route::get('create', [ListingController::class, 'create']);
-Route::post('create', [ListingController::class, 'store']);
+Route::get('create', [ListingController::class, 'create'])->middleware('auth');
+Route::post('create', [ListingController::class, 'store'])->middleware('auth');
 Route::get('listings/{listing}', [ListingController::class, 'show']);
-Route::get('listings/{listing}/edit', [ListingController::class, 'edit']);
-Route::put('listings/{listing}', [ListingController::class, 'update']);
-Route::delete('listings/{listing}', [ListingController::class, 'destroy']);
+Route::get('listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
+Route::put('listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
+Route::delete('listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+Route::get('register', [UserController::class, 'create'])->middleware('guest');
+Route::post('register', [UserController::class, 'store']);
+Route::get('logout', [UserController::class, 'logout'])->middleware('auth');
+Route::get('login', [UserController::class, 'login'])->name('login')->middleware('guest');
+Route::post('login', [UserController::class, 'loginNow']);
